+++ lib/omq/zmtp/routing/peer.rb
@@ -65,7 +65,7 @@ module OMQ
 
         def start_send_pump
           @send_pump_started = true
-          @tasks << @engine.parent_task.async(transient: true, annotation: "send pump") do
+          @tasks << @engine.spawn_pump_task(annotation: "send pump") do
             loop do
               @send_pump_idle = true
               batch = [@send_queue.dequeue]